+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+**BREAKING** Remove type states from the following drivers:
+
+- FlexPWM
+
+After construction, the type no longer tracks its pins or peripheral instances.
+To type check peripheral instances against pins, construct types using
+`with_pins`.
+
+**BREAKING** The FlexPWM refactor removes dedicated types for `Submodules` and
+`Outputs`. Users should program directly against the `Pwm` type.
+
 **BREAKING** Update `imxrt-usbd` to 0.3. Users are required to update to
 `usb-device` 0.3 and its compatible dependencies.